+++ tests/testthat/test-time_diff.R
@@ -76,22 +76,16 @@ testthat::test_that("time diff", {
   leap2 <- lubridate::dmy("28-02-2021")
   leap3 <- lubridate::dmy("01-03-2021")
   testthat::expect_true(floor(
-    time_diff(leap1, leap2, time_by = "year", as_period = FALSE,
-              time_type = "period")
-  ) == 1)
-  testthat::expect_true(floor(
-    time_diff(leap1, leap2, time_by = "year", as_period = TRUE,
-              time_type = "period")
-  ) == 0)
-  testthat::expect_true(floor(
-    time_diff(leap1, leap3, time_by = "year", as_period = FALSE,
+    time_diff(leap1, leap2, time_by = "year",
               time_type = "period")
   ) == 1)
   testthat::expect_true(
-    floor(
-      time_diff(leap1, leap3, time_by = "year", as_period = TRUE,
-                time_type = "period")
-    ) == 1)
+    age_years(leap1, leap2) == 0
+    )
+  testthat::expect_true(floor(
+    time_diff(leap1, leap3, time_by = lubridate::years(1))
+  ) == 1)
+  testthat::expect_true(age_years(leap1, leap3) == 1)
 
   # Test vectorization
   seq1 <- seq(-10, 10, 2)